+++ zed_components/src/zed_camera/include/zed_camera_component.hpp
@@ -687,6 +687,7 @@ private:
 
   // ----> Frame IDs
   bool mStaticTfPublished = false;
+  bool mStaticImuTfPublished = false;
   std::string mDepthFrameId;
   std::string mDepthOptFrameId;